JavascriptInterface

+++ src/org/jamruby/ext/JamActivity.mirah
@@ -6,8 +6,14 @@ import android.os.Handler
 import android.os.Looper
 import android.os.Environment
 import android.widget.Toast
+import android.widget.ScrollView
+import android.widget.LinearLayout
+import android.widget.LinearLayout.LayoutParams
+import android.widget.EditText
+import android.content.Context
 import android.util.Log
 import android.os.Process
+
 
 import java.io.File
 import java.io.InputStream
@@ -20,13 +26,38 @@ import org.jamruby.mruby.GC
 
 import Util
 import RubyObject
+
 import org.jamruby.ext.MainHandle
 import org.jamruby.ext.MainDispatch
 
+class DisplayError < LinearLayout
+  def initialize c:Context, msg:String, detail:String, backtrace:String[]
+    super(c)
+    
+    setOrientation LinearLayout.VERTICAL
+    
+    addView header = EditText.new(c), LayoutParams.new(LayoutParams.MATCH_PARENT, LayoutParams.WRAP_CONTENT, float(0.1));
+    addView trace  = EditText.new(c),  LayoutParams.new(LayoutParams.MATCH_PARENT, LayoutParams.MATCH_PARENT, float(1.0));
+
+    header.setMaxLines 1 
+    header.setHorizontallyScrolling(true);
+
+    trace.setHorizontallyScrolling(true);    
+
+    header.setText "Error: #{msg}"
+    
+    trace_msg = "#{detail}"
+    
+    backtrace.each do |m|
+      trace_msg += "\n#{String(m)}"
+    end
+    
+    trace.setText trace_msg
+  end
+end
+
 class JamActivity < Activity
   implements MainDispatch
-  
-  @@instance = JamActivity(nil)
   
   def onCreate state 
     super state
@@ -48,15 +79,25 @@ class JamActivity < Activity
     @main.core_libs.add "#{root}/mrblib/thread.mrb"        
     
     @_self_    = RubyObject(nil)
-  
-    @program = "#{root}/main.rb"
-    
-    init()
+    @create    = false
+    @program   = "#{root}/main.rb"
 
+    init
+    
     ol = ObjectList.create
     ol.addObj state
     
-    _self_.send "on_create", ol
+    r = _self_.send "on_create", ol
+    
+    if r.error
+      setContentView DisplayError.new(self, r.getErrorMessage, r.getErrorDetail, r.getErrorBacktrace)
+    else
+      @create = true
+    end
+  rescue => e
+    empty = String[1]
+    empty[0] = ""
+    setContentView DisplayError.new(self, "#{e}", "", empty)
   end
   
   def root
@@ -95,7 +136,7 @@ class JamActivity < Activity
     result  = main.jamruby.loadString("p $0; begin; ACTIVITY = Main.new; p ACTIVITY; ACTIVITY; rescue => e; p e; nil; end")
     Util.p result.toString
     @_self_ = RubyObject.new(main.jamruby.state, result)
-    Util.p _self_.send("to_s", ObjectList.create).toString
+    Util.p _self_.send("to_s", ObjectList.create).getResult.toString
     Util.p _self_.toString  
     Util.p "EMSG"  
   end
@@ -116,22 +157,15 @@ class JamActivity < Activity
     File.new(getFilesDir.toString+"/i").exists
   end
   
-  def selfCall fun:String
-    _self_.send fun, ObjectList.create
+  def selfCall(fun:String):boolean
+    r = _self_.send fun, ObjectList.create
+    r.getResult.isTrue
   end
   
   synchronized def selfCallArgv fun:String, ol:ObjectList
     s = _self_
     runOnUiThread do
       s.send fun, ol
-    end
-  end 
-  
-  synchronized def runProcOnMainThread(v:Value):void
-    state = main.jamruby.state
-    runOnUiThread do
-      i = MRuby.funcall(state, v, "__from_java__", 0)
-      MRuby.funcall(state, i, "__jam_call__", 1, MRuby.arrayNew(state))
     end
   end 
   
@@ -147,10 +181,10 @@ class JamActivity < Activity
   
   def onDestroy
     super
-    
-    selfCall("on_destroy")
-    
-    Process.killProcess(Process.myPid())
+    if !selfCall("on_destroy")
+      Util.p "destroy"
+      Process.killProcess(Process.myPid())
+    end
   end  
   
   def onResume
@@ -161,7 +195,13 @@ class JamActivity < Activity
   def onRestart
     super
     selfCall("on_restart")
-  end 
+  end
+  
+  def onBackPressed
+    if selfCall("on_back_pressed")
+      super
+    end
+  end    
   
   def install
     @didInstall = true
@@ -169,6 +209,7 @@ class JamActivity < Activity
     File.new("#{getFilesDir}/i").mkdir
     File.new("#{root}").mkdirs
     File.new("#{root}/mrblib").mkdirs  
+    File.new("#{root}/lib").mkdirs     
       
     am = getAssets();
     inputStream = am.open("main.rb");
@@ -185,6 +226,9 @@ class JamActivity < Activity
     
     inputStream = am.open("mrblib/activity.mrb");
     Util.createFileFromInputStream("#{root}/mrblib/activity.mrb", inputStream); 
+  
+    inputStream = am.open("lib/dynamic.rb");
+    Util.createFileFromInputStream("#{root}/lib/dynamic.rb", inputStream);   
     
     onInstall()
   end
